requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.1
django-allauth==65.0.2
django-appconf==1.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-imagekit==5.0.0
django-tz-detect==0.5.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.1
pilkit==3.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
urllib3==2.2.3
